+++ packages/sql-tag/src/factories/createSqlTokenSqlFragment.ts
@@ -9,6 +9,7 @@ import { createListSqlFragment } from '../sqlFragmentFactories/createListSqlFrag
 import { createQuerySqlFragment } from '../sqlFragmentFactories/createQuerySqlFragment';
 import { createTimestampSqlFragment } from '../sqlFragmentFactories/createTimestampSqlFragment';
 import { createUnnestSqlFragment } from '../sqlFragmentFactories/createUnnestSqlFragment';
+import { createUuidSqlFragment } from '../sqlFragmentFactories/createUuidSqlFragment';
 import {
   ArrayToken,
   BinaryToken,
@@ -22,6 +23,7 @@ import {
   QueryToken,
   TimestampToken,
   UnnestToken,
+  UuidToken,
 } from '../tokens';
 import { type SqlFragmentToken, type SqlToken as SqlTokenType } from '../types';
 import { UnexpectedStateError } from '@slonik/errors';
@@ -54,6 +56,8 @@ export const createSqlTokenSqlFragment = (
     return createTimestampSqlFragment(token, greatestParameterPosition);
   } else if (token.type === UnnestToken) {
     return createUnnestSqlFragment(token, greatestParameterPosition);
+  } else if (token.type === UuidToken) {
+    return createUuidSqlFragment(token, greatestParameterPosition);
   }
 
   throw new UnexpectedStateError('Unexpected token type.');