+++ packages/eslint-config/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @slonik/eslint-config
 
+## 46.3.0
+
 ## 46.2.0